requirements.txt
arch==6.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
fonttools==4.49.0
ibapi==10.27.1
idna==3.6
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.8.3
numpy==1.26.4
packaging==23.2
pandas==2.2.1
patsy==0.5.6
pillow==10.2.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2024.1
requests==2.31.0
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
statsmodels==0.14.1
threadpoolctl==3.3.0
tzdata==2024.1
urllib3==2.2.1
